Extract ChatInput component and further separation of concerns for AssistantPanel
